Added config.json to set up mailsettings email - so it can be used in the development stages. Using project's properties set there. Added ConfigurationBuilder to startup.cs specifying the AddJsonFile("confi.json"). Also Appcontroller has IConfigurationRoot added to the constructor that is injected in from the startup.cs. IConfigurationRoot  is part of ConfigurationBuilder. Inside the appcontroller Post method - use _config["MailSettings:ToAddress"] to retrieve the email address provided in the config.json.
In addition to thei sinsode the POst Method in AppController for contact - Added a means fo server validation of data. although the client side is validated the server side can also be validated and messages sent back to the client accordingly. Use if (ModelState.IsValid) to see if the data given is valid.
Can also add errors to the Modelstate -  ModelState.AddModelError("Email", "Dont support AOL.com - infact why would you be using that");

following a check to see if th eemail contains AOL.com the Model state can have the above added and thus an error thrown, the model state isnt therefore valid and the message is retruned to the client

Use ModelState.clear() to remove the data from the forms before returning view